requirements.txt
tensorflow>=2.10
streamlit>=1.28
opencv-python-headless
Pillow
numpy
pandas
matplotlib
scikit-learn
kaggle
